+++ app/partecipazioni.js
@@ -2,11 +2,12 @@ const express = require('express');
 const router = express.Router();
 const Partecipazione = require('./models/partecipazione.js');
 const Evento = require('./models/evento.js');
+const { tokenChecker } = require('./tokenChecker.js');
 
 router.post('/:id_evento', async (req,res) => {
     try{
         if(req.loggedUser.ruolo != 'cittadino'){
-            return res.status(403).json({ error: "Azione non permessa: la tipologia di utente non permette la proposta di un coupon"});
+            return res.status(403).json({ error: "Azione non permessa: la tipologia di utente non permette la proposta di un evento"});
         }
         
         let evento = await Evento.findById(req.params.id_evento);
@@ -26,19 +27,45 @@ router.post('/:id_evento', async (req,res) => {
     }
 });
 
-router.get('/:id_evento', async (req,res) => {
+// router.get('/:id_evento', async (req,res) => {
+//     try{
+//         if(req.loggedUser.ruolo != 'operatore_comunale'){
+//             return res.status(403).json({ error: "Azione non permessa: la tipologia di utente non permette la visualizzazione delle partecipazioni agli eventi"});
+//         }
+//         let filtro = {};
+//         filtro.id_evento = req.params.id_evento;
+//         if(!filtro){
+//             return res.status(400).json({ error: "Evento mancante"});
+//         }
+
+//         if(!Evento.findById(req.params.id_evento)){
+//             return res.status(404).json({ error: "Evento inesistente"});
+//         }
+
+//         let partecipazioni = await Partecipazione.find(filtro);
+//         res.status(200).json( partecipazioni.map(partecipazione => ({
+//             self: `/api/v1/partecipazioni/${partecipazione.id_evento}`,
+//             id_partecipazione: partecipazione._id,
+//             id_cittadino: req.loggedUser._id
+//         })));
+//     } catch (err) { 
+//         console.error(err);
+//         return res.status(500).json({ error: "Errore del server, riprova più tardi"});
+//     }
+// });
+
+router.get('', tokenChecker, async (req,res) => {
     try{
-        let filtro = {};
-        filtro.id_evento = req.params.id_evento;
-        if(!filtro){
-            return res.status(404).json({ error: "Evento non trovato"});
+        if(req.loggedUser.ruolo != 'cittadino'){
+            return res.status(403).json({ error: "Azione non permessa: la tipologia di utente non permette la visualizzazione delle partecipazioni agli eventi"});
         }
+        let filtro = {};
+        filtro.id_cittadino = req.loggedUser._id;
 
         let partecipazioni = await Partecipazione.find(filtro);
         res.status(200).json( partecipazioni.map(partecipazione => ({
-            self: `/api/v1/partecipazioni/${partecipazione.id_evento}`,
+            self: `/api/v1/partecipazioni`,
             id_partecipazione: partecipazione._id,
-            // id_evento: partecipazione.id_evento non credo sia necessario
             id_cittadino: req.loggedUser._id
         })));
     } catch (err) { 
@@ -46,14 +73,19 @@ router.get('/:id_evento', async (req,res) => {
         return res.status(500).json({ error: "Errore del server, riprova più tardi"});
     }
 });
-
-// Cancellazione coupon
-router.delete('/:id_evento', async (req,res) => {
+// Cancellazione partecipazione
+router.delete('/:id_evento', tokenChecker, async (req,res) => {
     try{
         const id_evento = req.params.id_evento;
         const evento = await Evento.findById(id_evento);
-        if(!evento) {
-            return res.status(404).json({ error: "Errore: evento inesistente" });
+        // if(!id_evento) {
+        //     return res.status(400).json({ error: "Evento mancante" });
+        // }
+        if(req.loggedUser.ruolo != 'cittadino'){
+            return res.status(403).json({ error: "Utente non autorizzato" });
+        }
+        if(!evento){
+            return res.status(404).json({ error: "Evento inesistente"});
         }
 
         const partecipazione = await Partecipazione.findOneAndDelete({
